requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
click==8.1.8
discord.py==2.5.2
Flask==3.1.0
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
propcache==0.3.1
python-dotenv==1.1.0
typing_extensions==4.13.2
Werkzeug==3.1.3
yarl==1.20.0